Fix ical file generation (#505)
Description of change

Fixes #467

As it stands, Outlook for Mac has a higher standard of compliance to the iCalendar standard, which our EventOccurencePage exports to. This higher standard was causing imported event for our .ics files to fail on calendar sync, and be deleted from the calendar.

To correct this three things have been changed:

    Additional fields have been added to the calendar (prodid and version), which are required for compliance.
    Additional fields have been added to the event (dtstamp and uid) which are required for compliance.
    The datetimes presented for dt_start and dt_end have been converted to UTC. The existing timezoned datetimes were being improperly formatted, which was what was actually breaking the imports.

Tests

An event was manually exported from an existing project to and .ics file, and imported into a freshly installed outlook for mac. This event was successfully added and editable.
Notes

While these changes correct the existing issue, the fact that the package used for handling iCalendar did not enforce compliance, and that its formatting broke with utc timezones make it a liability to use. In the future it should probably be removed for a simple formatted string/dict instead.