Hotfix timezones on ics export (#506)
In version 0.25.1, a fix was introduced which improved the compliance of the .ics exports of EventOccurrences. In an attempt to remove as many dependencies as possible, this change used python's builtin timezone.utc instead of pytz.utc as the timezone for dtstart and dtend.

Unfortunately, the iCalendar package is built around using pytz, and throws out non compliant output when not using pytz.

This reverts the change to using pytz, with the understanding that it will be removed when the iCalendar package is also removed.