Fix StreamForm submissions (#558)
StreamForms have been broken in 1.x, however they are seldomly used. I
noticed this after testing them with Wagtail 4, apparently, they were
broken by Wagtail 3.

In the event that anyone was using StreamForms with 1.x, I have included
a sample migration to fix the state of their database. However it is not
being included with coderedcms, because I wouldn't want to blindly
change `FormSubmission` objects (which is part of Wagtail, and used by
other forms as well). There is also a chance the site is using a custom
submission model, in which case the migration could not exist in
coderedcms. If the site is experiencing an error caused by this bug,
they should apply the migration themselves targeting the affected
models.