Bug: Correct length of classifier term length to 255 (#589)
#### Description of change
Changes the maximum length of the slug field of ClassifierTerm to have a
maximum length of 255, to match the name. This is also the maximum
length that several django backends allow for unique fields:
https://docs.djangoproject.com/en/4.2/ref/databases/#character-fields

#### Tests
Created a test project, ensured that a field with a name of a length
greater than 50 did not cause errors.

NOTES: This was believed to be fixed in 2.1.3, but that only corrected
the slug field on classifiers, not classifier terms.