Improve left/right/center image positions in rich text (#638)
Adds a bit of custom CSS to better align images in rich text in a way
that looks more natural/expected.
* Left and right images are capped at 40% width. This looks a bit more
natural, as large images could potentially fill the full width and don't
look left/right aligned.
* "Full width" image is not forced to 100%. Small images or icons look
too big and blurry. This is a bit more natural. Additionally, small
images which do not fill 100% are block positioned and centered.
* Headings in the rich text field will clear left/right floating
alignment. This looks a bit more natural and is an expected break after
an image.