Remove bootswatch themes (#640)
To support the custom navbar, and therefore simplify the settings, we
are removing the Bootswatch theme picker.

The theme dropdown was an ill-thought-out feature. We are now going in
the direction of simplifying the built-in navbar/settings, and
encouraging site builders to use custom sass and navbars. There's really
no reason a site admin should need to change the theme on the fly. That
is a nightmare that can remain relegated to WordPress.

"Themes" will be better served as a separate django app/package which
overrides templates. Using a dropdown of a limited number of pre-built
colorized Bootstrap builds is not really helpful in this regard.

For existing sites that want to continue using Bootswatch, it's pretty
easy to accomplish:

In your local project, edit
`website/templates/coderedcms/pages/base.html` and add/edit the
following block:

```jinja
{% block frontend_assets %}
<link rel="stylesheet" href="your bootswatch theme here">
{% endblock %}
```

The appropriate href for the `<link>` tag can be found here:
https://cdnjs.com/libraries/bootswatch