Use favicon instead of logo in wagtail admin (#650)
Starting with 4.0, sites with navbar settings disabled do not have the
ability to change the light/dark setting, which was being used in the
admin to put a light or dark background on the logo. Determining if the
logo is light or dark is a perpetual problem here. Because if this, the
default was light background, and a white logo is now unreadable.

To simplify it, let's switch to using the favicon. The favicon is much
more suitable for this kind of situation.