Bump dependencies and tests for Wagtail 6.3 Python 3.13 (#655)
No code changes required.

Note: since Wagtail 6.3 made some changes to the base media models, our
pro template is designed to work with Wagtail 6.3 out of the box. If you
are starting a new pro project with an earlier version of wagtail, it
will require clearing and re-generating the migrations. I think this is
an acceptable tradeoff rather than holding back on 6.3, since the
template is only used when creating new projects.

* Bump icalendar to 6.0
* Bump django-bootstrap5 to 24.3
* Update release pipeline to test on Python 3.13 and latest version of
wagtail.
* Remove python 3.8 from pipeline.