Fix logic for recaptcha v3 on error states (#698)
Certain error states in the recaptcha v3 API do not return a score.
However, we were defaulting to a "good" score and therefore the errors
were being let through.

This first checks if the request was successful. If so, it then compares
the score.

Fallbacks have also been changed from "good" states to sentinel values,
which should be helpful for debugging in production.